+++ planner/cardinality/BUILD.bazel
@@ -30,9 +30,11 @@ go_library(
         "//util/codec",
         "//util/collate",
         "//util/logutil",
+        "//util/mathutil",
         "//util/ranger",
         "//util/tracing",
         "@com_github_pingcap_errors//:errors",
+        "@com_github_pingcap_failpoint//:failpoint",
         "@org_golang_x_exp//maps",
         "@org_uber_go_zap//:zap",
     ],
