+++ planner/cardinality/row_count_column.go
@@ -20,6 +20,9 @@ import (
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/statistics"
 	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/util/codec"
+	"github.com/pingcap/tidb/util/collate"
+	"github.com/pingcap/tidb/util/mathutil"
 	"github.com/pingcap/tidb/util/ranger"
 )
 
@@ -29,6 +32,7 @@ func init() {
 	statistics.GetRowCountByIndexRanges = GetRowCountByIndexRanges
 	statistics.OutOfRangeEQSelectivity = outOfRangeEQSelectivity
 	statistics.GetEqualCondSelectivity = getEqualCondSelectivity
+	statistics.EqualRowCountOnColumn = EqualRowCountOnColumn
 }
 
 // GetRowCountByColumnRanges estimates the row count by a slice of Range.
@@ -62,7 +66,7 @@ func GetRowCountByColumnRanges(sctx sessionctx.Context, coll *statistics.HistCol
 			"Increase Factor", c.GetIncreaseFactor(coll.RealtimeCount),
 		)
 	}
-	result, err = c.GetColumnRowCount(sctx, colRanges, coll.RealtimeCount, coll.ModifyCount, false)
+	result, err = GetColumnRowCount(sctx, c, colRanges, coll.RealtimeCount, coll.ModifyCount, false)
 	if sc.EnableOptimizerCETrace {
 		CETraceRange(sctx, coll.PhysicalID, []string{c.Info.Name.O}, colRanges, "Column Stats", uint64(result))
 	}
@@ -107,9 +111,199 @@ func GetRowCountByIntColumnRanges(sctx sessionctx.Context, coll *statistics.Hist
 			"Increase Factor", c.GetIncreaseFactor(coll.RealtimeCount),
 		)
 	}
-	result, err = c.GetColumnRowCount(sctx, intRanges, coll.RealtimeCount, coll.ModifyCount, true)
+	result, err = GetColumnRowCount(sctx, c, intRanges, coll.RealtimeCount, coll.ModifyCount, true)
 	if sc.EnableOptimizerCETrace {
 		CETraceRange(sctx, coll.PhysicalID, []string{c.Info.Name.O}, intRanges, "Column Stats", uint64(result))
 	}
 	return result, errors.Trace(err)
+}
+
+// EqualRowCountOnColumn estimates the row count by a slice of Range and a Datum.
+func EqualRowCountOnColumn(sctx sessionctx.Context, c *statistics.Column, val types.Datum, encodedVal []byte, realtimeRowCount int64) (result float64, err error) {
+	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
+		debugtrace.EnterContextCommon(sctx)
+		debugtrace.RecordAnyValuesWithNames(sctx, "Value", val.String(), "Encoded", encodedVal)
+		defer func() {
+			debugtrace.RecordAnyValuesWithNames(sctx, "Result", result, "Error", err)
+			debugtrace.LeaveContextCommon(sctx)
+		}()
+	}
+	if val.IsNull() {
+		return float64(c.NullCount), nil
+	}
+	if c.StatsVer < statistics.Version2 {
+		// All the values are null.
+		if c.Histogram.Bounds.NumRows() == 0 {
+			return 0.0, nil
+		}
+		if c.Histogram.NDV > 0 && c.OutOfRange(val) {
+			return outOfRangeEQSelectivity(sctx, c.Histogram.NDV, realtimeRowCount, int64(c.TotalRowCount())) * c.TotalRowCount(), nil
+		}
+		if c.CMSketch != nil {
+			count, err := statistics.QueryValue(sctx, c.CMSketch, c.TopN, val)
+			return float64(count), errors.Trace(err)
+		}
+		histRowCount, _ := c.Histogram.EqualRowCount(sctx, val, false)
+		return histRowCount, nil
+	}
+
+	// Stats version == 2
+	// All the values are null.
+	if c.Histogram.Bounds.NumRows() == 0 && c.TopN.Num() == 0 {
+		return 0, nil
+	}
+	// 1. try to find this value in TopN
+	if c.TopN != nil {
+		rowcount, ok := c.TopN.QueryTopN(sctx, encodedVal)
+		if ok {
+			return float64(rowcount), nil
+		}
+	}
+	// 2. try to find this value in bucket.Repeat(the last value in every bucket)
+	histCnt, matched := c.Histogram.EqualRowCount(sctx, val, true)
+	if matched {
+		return histCnt, nil
+	}
+	// 3. use uniform distribution assumption for the rest (even when this value is not covered by the range of stats)
+	histNDV := float64(c.Histogram.NDV - int64(c.TopN.Num()))
+	if histNDV <= 0 {
+		return 0, nil
+	}
+	return c.Histogram.NotNullCount() / histNDV, nil
+}
+
+// GetColumnRowCount estimates the row count by a slice of Range.
+func GetColumnRowCount(sctx sessionctx.Context, c *statistics.Column, ranges []*ranger.Range, realtimeRowCount, modifyCount int64, pkIsHandle bool) (float64, error) {
+	sc := sctx.GetSessionVars().StmtCtx
+	debugTrace := sc.EnableOptimizerDebugTrace
+	if debugTrace {
+		debugtrace.EnterContextCommon(sctx)
+		defer debugtrace.LeaveContextCommon(sctx)
+	}
+	var rowCount float64
+	for _, rg := range ranges {
+		highVal := *rg.HighVal[0].Clone()
+		lowVal := *rg.LowVal[0].Clone()
+		if highVal.Kind() == types.KindString {
+			highVal.SetBytes(collate.GetCollator(highVal.Collation()).Key(highVal.GetString()))
+		}
+		if lowVal.Kind() == types.KindString {
+			lowVal.SetBytes(collate.GetCollator(lowVal.Collation()).Key(lowVal.GetString()))
+		}
+		cmp, err := lowVal.Compare(sc, &highVal, collate.GetBinaryCollator())
+		if err != nil {
+			return 0, errors.Trace(err)
+		}
+		lowEncoded, err := codec.EncodeKey(sc, nil, lowVal)
+		if err != nil {
+			return 0, err
+		}
+		highEncoded, err := codec.EncodeKey(sc, nil, highVal)
+		if err != nil {
+			return 0, err
+		}
+		if debugTrace {
+			statistics.DebugTraceStartEstimateRange(sctx, rg, lowEncoded, highEncoded, rowCount)
+		}
+		if cmp == 0 {
+			// case 1: it's a point
+			if !rg.LowExclude && !rg.HighExclude {
+				// In this case, the row count is at most 1.
+				if pkIsHandle {
+					rowCount++
+					if debugTrace {
+						statistics.DebugTraceEndEstimateRange(sctx, 1, statistics.DebugTraceUniquePoint)
+					}
+					continue
+				}
+				var cnt float64
+				cnt, err = EqualRowCountOnColumn(sctx, c, lowVal, lowEncoded, realtimeRowCount)
+				if err != nil {
+					return 0, errors.Trace(err)
+				}
+				// If the current table row count has changed, we should scale the row count accordingly.
+				cnt *= c.GetIncreaseFactor(realtimeRowCount)
+				rowCount += cnt
+				if debugTrace {
+					statistics.DebugTraceEndEstimateRange(sctx, cnt, statistics.DebugTracePoint)
+				}
+			}
+			continue
+		}
+		// In stats ver 1, we use CM Sketch to estimate row count for point condition, which is more accurate.
+		// So for the small range, we convert it to points.
+		if c.StatsVer < 2 {
+			rangeVals := statistics.EnumRangeValues(lowVal, highVal, rg.LowExclude, rg.HighExclude)
+
+			// case 2: it's a small range && using ver1 stats
+			if rangeVals != nil {
+				for _, val := range rangeVals {
+					cnt, err := EqualRowCountOnColumn(sctx, c, val, lowEncoded, realtimeRowCount)
+					if err != nil {
+						return 0, err
+					}
+					// If the current table row count has changed, we should scale the row count accordingly.
+					cnt *= c.GetIncreaseFactor(realtimeRowCount)
+					if debugTrace {
+						statistics.DebugTraceEndEstimateRange(sctx, cnt, statistics.DebugTraceVer1SmallRange)
+					}
+					rowCount += cnt
+				}
+
+				continue
+			}
+		}
+
+		// case 3: it's an interval
+		cnt := BetweenRowCountOnColumn(sctx, c, lowVal, highVal, lowEncoded, highEncoded)
+		// `betweenRowCount` returns count for [l, h) range, we adjust cnt for boundaries here.
+		// Note that, `cnt` does not include null values, we need specially handle cases
+		//   where null is the lower bound.
+		// And because we use (2, MaxValue] to represent expressions like a > 2 and use [MinNotNull, 3) to represent
+		//   expressions like b < 3, we need to exclude the special values.
+		if rg.LowExclude && !lowVal.IsNull() && lowVal.Kind() != types.KindMaxValue && lowVal.Kind() != types.KindMinNotNull {
+			lowCnt, err := EqualRowCountOnColumn(sctx, c, lowVal, lowEncoded, realtimeRowCount)
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+			cnt -= lowCnt
+			cnt = mathutil.Clamp(cnt, 0, c.NotNullCount())
+		}
+		if !rg.LowExclude && lowVal.IsNull() {
+			cnt += float64(c.NullCount)
+		}
+		if !rg.HighExclude && highVal.Kind() != types.KindMaxValue && highVal.Kind() != types.KindMinNotNull {
+			highCnt, err := EqualRowCountOnColumn(sctx, c, highVal, highEncoded, realtimeRowCount)
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+			cnt += highCnt
+		}
+
+		cnt = mathutil.Clamp(cnt, 0, c.TotalRowCount())
+
+		// If the current table row count has changed, we should scale the row count accordingly.
+		cnt *= c.GetIncreaseFactor(realtimeRowCount)
+
+		// handling the out-of-range part
+		if (c.OutOfRange(lowVal) && !lowVal.IsNull()) || c.OutOfRange(highVal) {
+			cnt += c.Histogram.OutOfRangeRowCount(sctx, &lowVal, &highVal, modifyCount)
+		}
+
+		if debugTrace {
+			statistics.DebugTraceEndEstimateRange(sctx, cnt, statistics.DebugTraceRange)
+		}
+		rowCount += cnt
+	}
+	rowCount = mathutil.Clamp(rowCount, 0, float64(realtimeRowCount))
+	return rowCount, nil
+}
+
+// BetweenRowCountOnColumn estimates the row count for interval [l, r).
+func BetweenRowCountOnColumn(sctx sessionctx.Context, c *statistics.Column, l, r types.Datum, lowEncoded, highEncoded []byte) float64 {
+	histBetweenCnt := c.Histogram.BetweenRowCount(sctx, l, r)
+	if c.StatsVer <= statistics.Version1 {
+		return histBetweenCnt
+	}
+	return float64(c.TopN.BetweenCount(sctx, lowEncoded, highEncoded)) + histBetweenCnt
 }
