+++ planner/cardinality/row_count_index.go
@@ -15,11 +15,22 @@
 package cardinality
 
 import (
+	"bytes"
 	"github.com/pingcap/errors"
+	"github.com/pingcap/failpoint"
+	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/planner/util/debugtrace"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/statistics"
+	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/util/chunk"
+	"github.com/pingcap/tidb/util/codec"
+	"github.com/pingcap/tidb/util/collate"
+	"github.com/pingcap/tidb/util/mathutil"
 	"github.com/pingcap/tidb/util/ranger"
+	"math"
+	"slices"
+	"strings"
 )
 
 // GetRowCountByIndexRanges estimates the row count by a slice of Range.
@@ -64,12 +75,430 @@ func GetRowCountByIndexRanges(sctx sessionctx.Context, coll *statistics.HistColl
 		)
 	}
 	if idx.CMSketch != nil && idx.StatsVer == statistics.Version1 {
-		result, err = coll.GetIndexRowCountForStatsV1(sctx, idxID, indexRanges)
+		result, err = GetIndexRowCountForStatsV1(sctx, coll, idxID, indexRanges)
 	} else {
-		result, err = idx.GetIndexRowCountForStatsV2(sctx, coll, indexRanges, coll.RealtimeCount, coll.ModifyCount)
+		result, err = GetIndexRowCountForStatsV2(sctx, idx, coll, indexRanges, coll.RealtimeCount, coll.ModifyCount)
 	}
 	if sc.EnableOptimizerCETrace {
 		CETraceRange(sctx, coll.PhysicalID, colNames, indexRanges, "Index Stats", uint64(result))
 	}
 	return result, errors.Trace(err)
+}
+
+// GetIndexRowCountForStatsV1 estimates the row count by the given range.
+func GetIndexRowCountForStatsV1(sctx sessionctx.Context, coll *statistics.HistColl, idxID int64, indexRanges []*ranger.Range) (float64, error) {
+	sc := sctx.GetSessionVars().StmtCtx
+	debugTrace := sc.EnableOptimizerDebugTrace
+	if debugTrace {
+		debugtrace.EnterContextCommon(sctx)
+		defer debugtrace.LeaveContextCommon(sctx)
+	}
+	idx := coll.Indices[idxID]
+	totalCount := float64(0)
+	for _, ran := range indexRanges {
+		if debugTrace {
+			statistics.DebugTraceStartEstimateRange(sctx, ran, nil, nil, totalCount)
+		}
+		rangePosition := statistics.GetOrdinalOfRangeCond(sc, ran)
+		var rangeVals []types.Datum
+		// Try to enum the last range values.
+		if rangePosition != len(ran.LowVal) {
+			rangeVals = statistics.EnumRangeValues(ran.LowVal[rangePosition], ran.HighVal[rangePosition], ran.LowExclude, ran.HighExclude)
+			if rangeVals != nil {
+				rangePosition++
+			}
+		}
+		// If first one is range, just use the previous way to estimate; if it is [NULL, NULL] range
+		// on single-column index, use previous way as well, because CMSketch does not contain null
+		// values in this case.
+		if rangePosition == 0 || isSingleColIdxNullRange(idx, ran) {
+			count, err := GetIndexRowCountForStatsV2(sctx, idx, nil, []*ranger.Range{ran}, coll.RealtimeCount, coll.ModifyCount)
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+			if debugTrace {
+				statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTraceRange)
+			}
+			totalCount += count
+			continue
+		}
+		var selectivity float64
+		// use CM Sketch to estimate the equal conditions
+		if rangeVals == nil {
+			bytes, err := codec.EncodeKey(sc, nil, ran.LowVal[:rangePosition]...)
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+			selectivity, err = statistics.GetEqualCondSelectivity(sctx, coll, idx, bytes, rangePosition, ran)
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+		} else {
+			bytes, err := codec.EncodeKey(sc, nil, ran.LowVal[:rangePosition-1]...)
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+			prefixLen := len(bytes)
+			for _, val := range rangeVals {
+				bytes = bytes[:prefixLen]
+				bytes, err = codec.EncodeKey(sc, bytes, val)
+				if err != nil {
+					return 0, err
+				}
+				res, err := statistics.GetEqualCondSelectivity(sctx, coll, idx, bytes, rangePosition, ran)
+				if err != nil {
+					return 0, errors.Trace(err)
+				}
+				selectivity += res
+			}
+		}
+		// use histogram to estimate the range condition
+		if rangePosition != len(ran.LowVal) {
+			rang := ranger.Range{
+				LowVal:      []types.Datum{ran.LowVal[rangePosition]},
+				LowExclude:  ran.LowExclude,
+				HighVal:     []types.Datum{ran.HighVal[rangePosition]},
+				HighExclude: ran.HighExclude,
+				Collators:   []collate.Collator{ran.Collators[rangePosition]},
+			}
+			var count float64
+			var err error
+			colIDs := coll.Idx2ColumnIDs[idxID]
+			var colID int64
+			if rangePosition >= len(colIDs) {
+				colID = -1
+			} else {
+				colID = colIDs[rangePosition]
+			}
+			// prefer index stats over column stats
+			if idxIDs, ok := coll.ColID2IdxIDs[colID]; ok && len(idxIDs) > 0 {
+				idxID := idxIDs[0]
+				count, err = GetRowCountByIndexRanges(sctx, coll, idxID, []*ranger.Range{&rang})
+			} else {
+				count, err = GetRowCountByColumnRanges(sctx, coll, colID, []*ranger.Range{&rang})
+			}
+			if err != nil {
+				return 0, errors.Trace(err)
+			}
+			selectivity = selectivity * count / idx.TotalRowCount()
+		}
+		count := selectivity * idx.TotalRowCount()
+		if debugTrace {
+			statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTraceRange)
+		}
+		totalCount += count
+	}
+	if totalCount > idx.TotalRowCount() {
+		totalCount = idx.TotalRowCount()
+	}
+	return totalCount, nil
+}
+
+// isSingleColIdxNullRange checks if a range is [NULL, NULL] on a single-column index.
+func isSingleColIdxNullRange(idx *statistics.Index, ran *ranger.Range) bool {
+	if len(idx.Info.Columns) > 1 {
+		return false
+	}
+	l, h := ran.LowVal[0], ran.HighVal[0]
+	if l.IsNull() && h.IsNull() {
+		return true
+	}
+	return false
+}
+
+// GetIndexRowCountForStatsV2 returns the row count of the given ranges.
+// It uses the modifyCount to adjust the influence of modifications on the table.
+func GetIndexRowCountForStatsV2(sctx sessionctx.Context, idx *statistics.Index, coll *statistics.HistColl, indexRanges []*ranger.Range, realtimeRowCount, modifyCount int64) (float64, error) {
+	idx.CheckStats()
+	sc := sctx.GetSessionVars().StmtCtx
+	debugTrace := sc.EnableOptimizerDebugTrace
+	if debugTrace {
+		debugtrace.EnterContextCommon(sctx)
+		defer debugtrace.LeaveContextCommon(sctx)
+	}
+	totalCount := float64(0)
+	isSingleCol := len(idx.Info.Columns) == 1
+	for _, indexRange := range indexRanges {
+		var count float64
+		lb, err := codec.EncodeKey(sc, nil, indexRange.LowVal...)
+		if err != nil {
+			return 0, err
+		}
+		rb, err := codec.EncodeKey(sc, nil, indexRange.HighVal...)
+		if err != nil {
+			return 0, err
+		}
+		if debugTrace {
+			statistics.DebugTraceStartEstimateRange(sctx, indexRange, lb, rb, totalCount)
+		}
+		fullLen := len(indexRange.LowVal) == len(indexRange.HighVal) && len(indexRange.LowVal) == len(idx.Info.Columns)
+		if bytes.Equal(lb, rb) {
+			// case 1: it's a point
+			if indexRange.LowExclude || indexRange.HighExclude {
+				if debugTrace {
+					statistics.DebugTraceEndEstimateRange(sctx, 0, statistics.DebugTraceImpossible)
+				}
+				continue
+			}
+			if fullLen {
+				// At most 1 in this case.
+				if idx.Info.Unique {
+					totalCount++
+					if debugTrace {
+						statistics.DebugTraceEndEstimateRange(sctx, 1, statistics.DebugTraceUniquePoint)
+					}
+					continue
+				}
+				count = equalRowCountOnIndex(sctx, idx, lb, realtimeRowCount)
+				// If the current table row count has changed, we should scale the row count accordingly.
+				count *= idx.GetIncreaseFactor(realtimeRowCount)
+				if debugTrace {
+					statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTracePoint)
+				}
+				totalCount += count
+				continue
+			}
+		}
+
+		// case 2: it's an interval
+		// The final interval is [low, high)
+		if indexRange.LowExclude {
+			lb = kv.Key(lb).PrefixNext()
+		}
+		if !indexRange.HighExclude {
+			rb = kv.Key(rb).PrefixNext()
+		}
+		l := types.NewBytesDatum(lb)
+		r := types.NewBytesDatum(rb)
+		lowIsNull := bytes.Equal(lb, nullKeyBytes)
+		if isSingleCol && lowIsNull {
+			count += float64(idx.Histogram.NullCount)
+		}
+		expBackoffSuccess := false
+		// Due to the limitation of calcFraction and convertDatumToScalar, the histogram actually won't estimate anything.
+		// If the first column's range is point.
+		if rangePosition := statistics.GetOrdinalOfRangeCond(sc, indexRange); rangePosition > 0 && idx.StatsVer >= statistics.Version2 && coll != nil {
+			var expBackoffSel float64
+			expBackoffSel, expBackoffSuccess, err = expBackoffEstimation(sctx, idx, coll, indexRange)
+			if err != nil {
+				return 0, err
+			}
+			if expBackoffSuccess {
+				expBackoffCnt := expBackoffSel * idx.TotalRowCount()
+
+				upperLimit := expBackoffCnt
+				// Use the multi-column stats to calculate the max possible row count of [l, r)
+				if idx.Histogram.Len() > 0 {
+					_, lowerBkt, _, _ := idx.Histogram.LocateBucket(sctx, l)
+					_, upperBkt, _, _ := idx.Histogram.LocateBucket(sctx, r)
+					if debugTrace {
+						statistics.DebugTraceBuckets(sctx, &idx.Histogram, []int{lowerBkt - 1, upperBkt})
+					}
+					// Use Count of the Bucket before l as the lower bound.
+					preCount := float64(0)
+					if lowerBkt > 0 {
+						preCount = float64(idx.Histogram.Buckets[lowerBkt-1].Count)
+					}
+					// Use Count of the Bucket where r exists as the upper bound.
+					upperCnt := float64(idx.Histogram.Buckets[upperBkt].Count)
+					upperLimit = upperCnt - preCount
+					upperLimit += float64(idx.TopN.BetweenCount(sctx, lb, rb))
+				}
+
+				// If the result of exponential backoff strategy is larger than the result from multi-column stats,
+				// 	use the upper limit from multi-column histogram instead.
+				if expBackoffCnt > upperLimit {
+					expBackoffCnt = upperLimit
+				}
+				count += expBackoffCnt
+			}
+		}
+		if !expBackoffSuccess {
+			count += BetweenRowCountOnIndex(sctx, idx, l, r)
+		}
+
+		// If the current table row count has changed, we should scale the row count accordingly.
+		count *= idx.GetIncreaseFactor(realtimeRowCount)
+
+		// handling the out-of-range part
+		if (OutOfRangeOnIndex(idx, l) && !(isSingleCol && lowIsNull)) || OutOfRangeOnIndex(idx, r) {
+			count += idx.Histogram.OutOfRangeRowCount(sctx, &l, &r, modifyCount)
+		}
+
+		if debugTrace {
+			statistics.DebugTraceEndEstimateRange(sctx, count, statistics.DebugTraceRange)
+		}
+		totalCount += count
+	}
+	totalCount = mathutil.Clamp(totalCount, 0, float64(realtimeRowCount))
+	return totalCount, nil
+}
+
+var nullKeyBytes, _ = codec.EncodeKey(nil, nil, types.NewDatum(nil))
+
+func equalRowCountOnIndex(sctx sessionctx.Context, idx *statistics.Index, b []byte, realtimeRowCount int64) (result float64) {
+	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
+		debugtrace.EnterContextCommon(sctx)
+		debugtrace.RecordAnyValuesWithNames(sctx, "Encoded Value", b)
+		defer func() {
+			debugtrace.RecordAnyValuesWithNames(sctx, "Result", result)
+			debugtrace.LeaveContextCommon(sctx)
+		}()
+	}
+	if len(idx.Info.Columns) == 1 {
+		if bytes.Equal(b, nullKeyBytes) {
+			return float64(idx.Histogram.NullCount)
+		}
+	}
+	val := types.NewBytesDatum(b)
+	if idx.StatsVer < statistics.Version2 {
+		if idx.Histogram.NDV > 0 && OutOfRangeOnIndex(idx, val) {
+			return outOfRangeEQSelectivity(sctx, idx.Histogram.NDV, realtimeRowCount, int64(idx.TotalRowCount())) * idx.TotalRowCount()
+		}
+		if idx.CMSketch != nil {
+			return float64(idx.QueryBytes(sctx, b))
+		}
+		histRowCount, _ := idx.Histogram.EqualRowCount(sctx, val, false)
+		return histRowCount
+	}
+	// stats version == 2
+	// 1. try to find this value in TopN
+	if idx.TopN != nil {
+		count, found := idx.TopN.QueryTopN(sctx, b)
+		if found {
+			return float64(count)
+		}
+	}
+	// 2. try to find this value in bucket.Repeat(the last value in every bucket)
+	histCnt, matched := idx.Histogram.EqualRowCount(sctx, val, true)
+	if matched {
+		return histCnt
+	}
+	// 3. use uniform distribution assumption for the rest (even when this value is not covered by the range of stats)
+	histNDV := float64(idx.Histogram.NDV - int64(idx.TopN.Num()))
+	if histNDV <= 0 {
+		return 0
+	}
+	return idx.Histogram.NotNullCount() / histNDV
+}
+
+// expBackoffEstimation estimate the multi-col cases following the Exponential Backoff. See comment below for details.
+func expBackoffEstimation(sctx sessionctx.Context, idx *statistics.Index, coll *statistics.HistColl, indexRange *ranger.Range) (sel float64, success bool, err error) {
+	if sctx.GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
+		debugtrace.EnterContextCommon(sctx)
+		defer func() {
+			debugtrace.RecordAnyValuesWithNames(sctx,
+				"Result", sel,
+				"Success", success,
+				"error", err,
+			)
+			debugtrace.LeaveContextCommon(sctx)
+		}()
+	}
+	tmpRan := []*ranger.Range{
+		{
+			LowVal:    make([]types.Datum, 1),
+			HighVal:   make([]types.Datum, 1),
+			Collators: make([]collate.Collator, 1),
+		},
+	}
+	colsIDs := coll.Idx2ColumnIDs[idx.Histogram.ID]
+	singleColumnEstResults := make([]float64, 0, len(indexRange.LowVal))
+	// The following codes uses Exponential Backoff to reduce the impact of independent assumption. It works like:
+	//   1. Calc the selectivity of each column.
+	//   2. Sort them and choose the first 4 most selective filter and the corresponding selectivity is sel_1, sel_2, sel_3, sel_4 where i < j => sel_i < sel_j.
+	//   3. The final selectivity would be sel_1 * sel_2^{1/2} * sel_3^{1/4} * sel_4^{1/8}.
+	// This calculation reduced the independence assumption and can work well better than it.
+	for i := 0; i < len(indexRange.LowVal); i++ {
+		tmpRan[0].LowVal[0] = indexRange.LowVal[i]
+		tmpRan[0].HighVal[0] = indexRange.HighVal[i]
+		tmpRan[0].Collators[0] = indexRange.Collators[0]
+		if i == len(indexRange.LowVal)-1 {
+			tmpRan[0].LowExclude = indexRange.LowExclude
+			tmpRan[0].HighExclude = indexRange.HighExclude
+		}
+		colID := colsIDs[i]
+		var (
+			count      float64
+			err        error
+			foundStats bool
+		)
+		if col, ok := coll.Columns[colID]; ok && !col.IsInvalid(sctx, coll.Pseudo) {
+			foundStats = true
+			count, err = GetRowCountByColumnRanges(sctx, coll, colID, tmpRan)
+		}
+		if idxIDs, ok := coll.ColID2IdxIDs[colID]; ok && !foundStats && len(indexRange.LowVal) > 1 {
+			// Note the `len(indexRange.LowVal) > 1` condition here, it means we only recursively call
+			// `GetRowCountByIndexRanges()` when the input `indexRange` is a multi-column range. This
+			// check avoids infinite recursion.
+			for _, idxID := range idxIDs {
+				if idxID == idx.Histogram.ID {
+					continue
+				}
+				foundStats = true
+				count, err = GetRowCountByIndexRanges(sctx, coll, idxID, tmpRan)
+				if err == nil {
+					break
+				}
+			}
+		}
+		if !foundStats {
+			continue
+		}
+		if err != nil {
+			return 0, false, err
+		}
+		singleColumnEstResults = append(singleColumnEstResults, count)
+	}
+	// Sort them.
+	slices.Sort(singleColumnEstResults)
+	l := len(singleColumnEstResults)
+	// Convert the first 4 to selectivity results.
+	for i := 0; i < l && i < 4; i++ {
+		singleColumnEstResults[i] = singleColumnEstResults[i] / float64(coll.RealtimeCount)
+	}
+	failpoint.Inject("cleanEstResults", func() {
+		singleColumnEstResults = singleColumnEstResults[:0]
+		l = 0
+	})
+	if l == 1 {
+		return singleColumnEstResults[0], true, nil
+	} else if l == 2 {
+		return singleColumnEstResults[0] * math.Sqrt(singleColumnEstResults[1]), true, nil
+	} else if l == 3 {
+		return singleColumnEstResults[0] * math.Sqrt(singleColumnEstResults[1]) * math.Sqrt(math.Sqrt(singleColumnEstResults[2])), true, nil
+	} else if l == 0 {
+		return 0, false, nil
+	}
+	return singleColumnEstResults[0] * math.Sqrt(singleColumnEstResults[1]) * math.Sqrt(math.Sqrt(singleColumnEstResults[2])) * math.Sqrt(math.Sqrt(math.Sqrt(singleColumnEstResults[3]))), true, nil
+}
+
+// OutOfRangeOnIndex checks if the datum is out of the range.
+func OutOfRangeOnIndex(idx *statistics.Index, val types.Datum) bool {
+	if !idx.Histogram.OutOfRange(val) {
+		return false
+	}
+	if idx.Histogram.Len() > 0 && matchPrefix(idx.Histogram.Bounds.GetRow(0), 0, &val) {
+		return false
+	}
+	return true
+}
+
+// matchPrefix checks whether ad is the prefix of value
+func matchPrefix(row chunk.Row, colIdx int, ad *types.Datum) bool {
+	switch ad.Kind() {
+	case types.KindString, types.KindBytes, types.KindBinaryLiteral, types.KindMysqlBit:
+		return strings.HasPrefix(row.GetString(colIdx), ad.GetString())
+	}
+	return false
+}
+
+// BetweenRowCountOnIndex estimates the row count for interval [l, r).
+// The input sctx is just for debug trace, you can pass nil safely if that's not needed.
+func BetweenRowCountOnIndex(sctx sessionctx.Context, idx *statistics.Index, l, r types.Datum) float64 {
+	histBetweenCnt := idx.Histogram.BetweenRowCount(sctx, l, r)
+	if idx.StatsVer == statistics.Version1 {
+		return histBetweenCnt
+	}
+	return float64(idx.TopN.BetweenCount(sctx, l.GetBytes(), r.GetBytes())) + histBetweenCnt
 }
